+++ index.html
@@ -10,6 +10,7 @@
 </head>
 
 <body>
+  <script src="/js/main.js" type="module"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -231,7 +232,5 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
 </body>
 </html>
